requirements.txt
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
et-xmlfile==1.1.0
filelock==3.7.1
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.4.0
gspread-dataframe==3.3.0
gspread-formatting==1.0.6
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.9.0
numpy==1.22.4
oauthlib==3.2.0
openpyxl==3.0.10
outcome==1.1.0
pandas==1.4.2
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
Scrapy==2.6.1
scrapy-xlsx==0.1.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tldextract==3.3.0
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.10
w3lib==1.22.0
wsproto==1.1.0
zope.interface==5.4.0
